Move RevisionMixin to revisions submodule
